Initial Commit for PmodOLED driver
THis commit includes the driver for PmodOLED device (compatible with on-board
OLED for ZED Rev.B as well). ZED on-board OLED device is added into ZED default
device tree. PmodOLED is selected in digilent_zed_defconfig.

There is a bug when PmodOLED is compiled as loadable module. The bug may
cause kernel panic after inserting to and removing from the kernel for several times.