axidma_of:Read correct channel ID from DT
The original author of the driver read the channel ID form the xlnx,device-id of the dma channel-node within the DT. This field however represents the device ID of the DMA rather then an individual ID for the channel.

The channel ID is codded in to the dmas property of the chardev in the DT as the first argument.
Ex:

dmas = <&axi_dma_0 1>;

in this case the channel ID is 1, and a RX channel. The TX channel ID is 0 and RX channel ID is 1.